refactor(zone.js): move patchClass() code to legacy bundle
